fix: curly bracket interpreted as markdown attribute
Curly bracketes right after starting a markdown code blocks might be interpreted as
an markdown attribute. If this is the case, adding a new line right after the code
block starts fixes it. This occured with
https://github.com/bitcoin/bitcoin/issues/12355#issue-294394225 (where
GitHub does not correctly render the first code block)

Workaround for https://github.com/gohugoio/hugo/issues/10835